requirements.txt
loggly-python-handler

